+++ src/index.js
@@ -1,34 +1,61 @@
 import mixed from './mixed';
 import bool from './boolean';
+import string from './string';
+import number from './number';
+import date from './date';
+import object from './object';
+import array from './array';
 import Ref from './Reference';
 import Lazy from './Lazy';
+import ValidationError from './ValidationError';
+import reach from './util/reach';
 import isSchema from './util/isSchema';
 
-export default {
-  mixed:   mixed,
-  string:  require('./string'),
-  number:  require('./number'),
-  boolean: bool,
-  bool:    bool,
-  date:    require('./date'),
-  object:  require('./object'),
-  array:   require('./array'),
+let boolean = bool;
+let ref =(key, options) => new Ref(key, options);
 
-  reach: require('./util/reach'),
+let lazy =(fn) => new Lazy(fn);
 
-  ValidationError: require('./ValidationError'),
-  ref: (key, options) => new Ref(key, options),
-  lazy: (fn) => new Lazy(fn),
+function addMethod(schemaType, name, fn) {
+  if (!schemaType || !isSchema(schemaType.prototype))
+    throw new TypeError('You must provide a yup schema constructor function')
 
-  isSchema,
+  if (typeof name !== 'string') throw new TypeError('A Method name must be provided')
+  if (typeof fn !== 'function') throw new TypeError('Method function must be provided')
 
-  addMethod(schemaType, name, fn) {
-    if ( !schemaType || !isSchema(schemaType.prototype))
-      throw new TypeError('You must provide a yup schema constructor function')
-
-    if ( typeof name !== 'string') throw new TypeError('A Method name must be provided')
-    if ( typeof fn !== 'function') throw new TypeError('Method function must be provided')
-
-    schemaType.prototype[name] = fn
-  }
+  schemaType.prototype[name] = fn
 }
+
+export {
+  mixed,
+  string,
+  number,
+  bool,
+  boolean,
+  date,
+  object,
+  array,
+  ref,
+  lazy,
+  reach,
+  isSchema,
+  addMethod,
+  ValidationError,
+}
+
+export default {
+  mixed,
+  string,
+  number,
+  bool,
+  boolean,
+  date,
+  object,
+  array,
+  ref,
+  lazy,
+  reach,
+  isSchema,
+  addMethod,
+  ValidationError,
+};