+++ src/util/isAbsent.js
@@ -1,2 +1,2 @@
 
-module.exports = value => value == null
+export default value => value == null