+++ src/util/isodate.js
@@ -7,7 +7,7 @@
 //              1 YYYY                 2 MM        3 DD              4 HH     5 mm        6 ss            7 msec         8 Z 9 ±    10 tzHH    11 tzmm
 var isoReg = /^(\d{4}|[+\-]\d{6})(?:-?(\d{2})(?:-?(\d{2}))?)?(?:[ T]?(\d{2}):?(\d{2})(?::?(\d{2})(?:[,\.](\d{1,}))?)?(?:(Z)|([+\-])(\d{2})(?::?(\d{2}))?)?)?$/
 
-module.exports = function parseIsoDate(date) {
+export default function parseIsoDate(date) {
   var numericKeys = [ 1, 4, 5, 6, 7, 10, 11 ]
     , minutesOffset = 0
     , timestamp, struct;
@@ -32,7 +32,7 @@ module.exports = function parseIsoDate(date) {
       if (struct[8] !== 'Z' && struct[9] !== undefined) {
         minutesOffset = struct[10] * 60 + struct[11];
 
-        if (struct[9] === '+') 
+        if (struct[9] === '+')
           minutesOffset = 0 - minutesOffset;
       }