+++ test/object.js
@@ -1,5 +1,5 @@
-
 import Promise from 'promise/src/es6-extensions';
+
 import {
     mixed, string, date, number
   , bool, array, object, ref, lazy, reach
@@ -8,7 +8,7 @@ import {
 describe('Object types', function(){
 
   describe('casting', ()=> {
-    it ('should parse json strings', () => {
+    it('should parse json strings', () => {
       object({ hello: number() })
         .cast('{ \"hello\": \"5\" }')
         .should.eql({
@@ -16,12 +16,12 @@ describe('Object types', function(){
         })
     })
 
-    it ('should return null for failed casts', () => {
+    it('should return null for failed casts', () => {
       expect(
         object().cast('dfhdfh', { assert: false })).to.equal(null)
     })
 
-    it ('should recursively cast fields', () => {
+    it('should recursively cast fields', () => {
       var obj = {
         num: '5',
         str: 'hello',
@@ -75,16 +75,17 @@ describe('Object types', function(){
     })
 
     it('should run validations recursively', async () => {
-      let error = await inst.validate(obj).should.be.rejected;
+      let error = await inst.validate(obj).should.be.rejected();
 
       error.errors.length.should.equal(1)
       error.errors[0].should.contain('nested.str')
 
       obj.arr[1] = 8
 
-      await inst.isValid().should.eventually.equal(true),
+      await inst.isValid().should.eventually().equal(true)
 
-      error = await inst.validate(obj).should.be.rejected
+      error = await inst.validate(obj).should.be.rejected()
+
       error.errors[0].should.contain('arr[1]')
     })
 
@@ -99,7 +100,7 @@ describe('Object types', function(){
 
       value.field.should.equal('5')
 
-      castSpy.should.have.been.calledOnce
+      castSpy.should.have.been.calledOnce()
 
       string.prototype._cast.restore()
     })
@@ -110,7 +111,7 @@ describe('Object types', function(){
       })
       .strict()
 
-      let err = await inst.validate({ field: 5 }).should.be.rejected
+      let err = await inst.validate({ field: 5 }).should.be.rejected()
 
       err.message.should.match(/must be a `string` type/)
     })
@@ -120,7 +121,7 @@ describe('Object types', function(){
         field: number().strict()
       })
 
-      let err = await inst.validate({ field: '5' }).should.be.rejected
+      let err = await inst.validate({ field: '5' }).should.be.rejected()
 
       err.message.should.match(/must be a `number` type/)
 
@@ -132,7 +133,7 @@ describe('Object types', function(){
           .integer()
       })
       .validate({ port: 'asdad' })
-      .should.be.rejected
+      .should.be.rejected()
     })
 
     it('should handle custom validation', async function(){
@@ -142,7 +143,7 @@ describe('Object types', function(){
       })
       .test('test', '${path} oops', () => false)
 
-      let err = await inst.validate({}).should.be.rejected
+      let err = await inst.validate({}).should.be.rejected()
 
       err.errors[0].should.equal('this oops')
     })
@@ -214,13 +215,13 @@ describe('Object types', function(){
       })
     })
 
-    it ('should expand objects by default', () => {
+    it('should expand objects by default', () => {
       objSchema.default().should.eql({
         nest: { str: 'hi' }
       })
     })
 
-    it ('should accept a user provided default', () => {
+    it('should accept a user provided default', () => {
       objSchema = objSchema.default({ boom: 'hi'})
 
       objSchema.default().should.eql({
@@ -228,7 +229,7 @@ describe('Object types', function(){
       })
     })
 
-    it ('should add empty keys when sub schema has no default', () => {
+    it('should add empty keys when sub schema has no default', () => {
       object({
         str: string(),
         nest: object({ str: string() })
@@ -240,7 +241,7 @@ describe('Object types', function(){
       })
     })
 
-    it ('should create defaults for missing object fields', () => {
+    it('should create defaults for missing object fields', () => {
 
       object({
         prop: mixed(),
@@ -263,10 +264,10 @@ describe('Object types', function(){
 
     return Promise.all([
 
-      inst.isValid({}).should.eventually.equal(true),
+      inst.isValid({}).should.eventually().equal(true),
 
       inst.shape({ prop: mixed().required() })
-        .isValid({}).should.eventually.equal(false)
+        .isValid({}).should.eventually().equal(false)
     ])
   })
 
@@ -279,14 +280,14 @@ describe('Object types', function(){
     return Promise.all([
       inst
         .noUnknown('hi')
-        .validate({ extra: 'field' }, { strict: true }).should.be.rejected
+        .validate({ extra: 'field' }, { strict: true }).should.be.rejected()
           .then(function(err){
             err.errors[0].should.equal('hi')
           }),
 
       inst
         .noUnknown()
-        .validate({ extra: 'field' }, { strict: true }).should.be.rejected
+        .validate({ extra: 'field' }, { strict: true }).should.be.rejected()
           .then(function(err){
             err.errors[0].should.be.a('string')
           })
@@ -461,7 +462,7 @@ describe('Object types', function(){
 
     return Promise.all([
       inst
-        .validate({ nest: { str: '' } }).should.be.rejected
+        .validate({ nest: { str: '' } }).should.be.rejected()
         .then(function(err) {
           err.value.should.eql({ nest: { str: '' }  })
           err.errors.length.should.equal(1)
@@ -475,7 +476,7 @@ describe('Object types', function(){
           { nest: { str: '' } },
           { abortEarly: false }
         )
-        .should.be.rejected
+        .should.be.rejected()
         .then(function(err) {
           err.value.should.eql({ nest: { str: '' } })
           err.errors.length.should.equal(2)
@@ -487,14 +488,16 @@ describe('Object types', function(){
   it('should sort errors by insertion order', async () => {
     var inst = object({
       foo: string().test('foo', function() {
-        return new Promise(resolve => setTimeout(() => resolve(false), 10))
+        return new Promise(resolve => {
+          setTimeout(() => resolve(false), 10)
+        })
       }),
       bar: string().required()
     })
 
-    let err = await inst.validate(
-      { foo: 'foo' },
-      { abortEarly: false }).should.rejected;
+    let err = await inst
+      .validate({ foo: 'foo' }, { abortEarly: false })
+      .should.rejected();
 
     err.errors.should.eql([
       'foo is invalid',
@@ -514,13 +517,13 @@ describe('Object types', function(){
 
     return Promise.all([
       inst
-      .validate(val, { abortEarly: false }).should.be.rejected
+      .validate(val, { abortEarly: false }).should.be.rejected()
       .then(function(err){
         err.errors.length.should.equal(2)
       }),
 
       inst
-        .validate(val, { abortEarly: false, recursive: false }).should.be.rejected
+        .validate(val, { abortEarly: false, recursive: false }).should.be.rejected()
         .then(function(err){
           err.errors.length.should.equal(1)
           err.errors.should.eql(['oops'])
@@ -566,21 +569,21 @@ describe('Object types', function(){
         })
 
     return Promise.all([
-      inst.isValid({ stats: { isBig: true }, rand: 5, noteDate: 7, other: 4 }).should.eventually.equal(false),
-      inst.isValid({ stats: { isBig: true }, noteDate: 1, other: 4 }).should.eventually.equal(false),
+      inst.isValid({ stats: { isBig: true }, rand: 5, noteDate: 7, other: 4 }).should.eventually().equal(false),
+      inst.isValid({ stats: { isBig: true }, noteDate: 1, other: 4 }).should.eventually().equal(false),
 
-      inst.isValid({ stats: { isBig: true }, noteDate: 7, other: 6 }).should.eventually.equal(true),
-      inst.isValid({ stats: { isBig: true }, noteDate: 7, other: 4 }).should.eventually.equal(false),
+      inst.isValid({ stats: { isBig: true }, noteDate: 7, other: 6 }).should.eventually().equal(true),
+      inst.isValid({ stats: { isBig: true }, noteDate: 7, other: 4 }).should.eventually().equal(false),
 
-      inst.isValid({ stats: { isBig: false }, noteDate: 4, other: 4 }).should.eventually.equal(true),
+      inst.isValid({ stats: { isBig: false }, noteDate: 4, other: 4 }).should.eventually().equal(true),
 
-      inst.isValid({ stats: { isBig: true }, noteDate: 1, other: 4 }).should.eventually.equal(false),
-      inst.isValid({ stats: { isBig: true }, noteDate: 6, other: 4 }).should.eventually.equal(true)
+      inst.isValid({ stats: { isBig: true }, noteDate: 1, other: 4 }).should.eventually().equal(false),
+      inst.isValid({ stats: { isBig: true }, noteDate: 6, other: 4 }).should.eventually().equal(true)
     ])
   })
 
   it('should allow opt out of topo sort on specific edges', function(){
-    !(function() {
+    (function() {
       object().shape({
           orgID: number()
             .when('location', function(v){ if (v == null) return this.required() }),
@@ -589,7 +592,7 @@ describe('Object types', function(){
         })
     }).should.throw('Cyclic dependency: "location"')
 
-    !(function() {
+    ;(function() {
       object().shape({
           orgID: number()
             .when('location', function(v){ if (v == null) return this.required() }),
@@ -623,22 +626,22 @@ describe('Object types', function(){
         })
 
     return Promise.all([
-      inst.validate({ stats: undefined, other: true }).should.be.rejected
+      inst.validate({ stats: undefined, other: true }).should.be.rejected()
         .then(function(err){
           err.errors[0].should.contain('required')
         }),
 
-      inst.validate({ stats: { isBig: true, count: 3 }, other: true }).should.be.rejected
+      inst.validate({ stats: { isBig: true, count: 3 }, other: true }).should.be.rejected()
         .then(function(err){
           err.errors[0].should.contain('must be greater than or equal to 5')
         }),
 
-      inst.validate({ stats: { isBig: true, count: 10 }, other: true }).should.be.fulfilled
+      inst.validate({ stats: { isBig: true, count: 10 }, other: true }).should.be.fulfilled()
         .then(function(value){
           value.should.deep.equal({ stats: { isBig: true, count: 10 }, other: true })
         }),
 
-      countSchema.validate(10, { context: { isBig: true } }).should.be.fulfilled
+      countSchema.validate(10, { context: { isBig: true } }).should.be.fulfilled()
         .then(function(value){
           value.should.deep.equal(10)
         })
@@ -651,7 +654,7 @@ describe('Object types', function(){
           caseStatus: number(),
           hiJohn: number()
         })
-        .camelcase()
+        .camelCase()
 
     inst.cast({ CON_STAT: 5, CaseStatus: 6, 'hi john': 4 })
       .should.eql({ conStat: 5, caseStatus: 6, hiJohn: 4 })
@@ -661,14 +664,14 @@ describe('Object types', function(){
       .cast(null)).to.equal(null)
   })
 
-  it('should camelCase with leading underscore', function(){
-    var inst = object().camelcase()
-
-    inst
-      .cast({ CON_STAT: 5, __isNew: true, __IS_FUN: true })
-      .should
-      .eql({ conStat: 5, __isNew: true, __isFun: true })
-  })
+  // it('should camelCase with leading underscore', function(){
+  //   var inst = object().camelCase()
+  //
+  //   inst
+  //     .cast({ CON_STAT: 5, __isNew: true, __IS_FUN: true })
+  //     .should
+  //     .eql({ conStat: 5, __isNew: true, __isFun: true })
+  // })
 
   it('should CONSTANT_CASE keys', function(){
     var inst = object().shape({
@@ -676,7 +679,7 @@ describe('Object types', function(){
           CASE_STATUS: number(),
           HI_JOHN: number()
         })
-        .constantcase()
+        .constantCase()
 
     inst.cast({ conStat: 5, CaseStatus: 6, 'hi john': 4 })
       .should.eql({ CON_STAT: 5, CASE_STATUS: 6, HI_JOHN: 4 })
@@ -685,6 +688,4 @@ describe('Object types', function(){
       .nullable()
       .cast(null)).to.equal(null)
   })
-
-
 })